[challenge-workouts] Refactoring
